requirements.txt
aioredis==1.3.1
amqp==5.0.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
billiard==3.6.3.0
braintree==4.5.0
cairocffi==1.2.0
CairoSVG==2.5.0
celery==5.0.2
certifi==2020.11.8
cffi==1.14.3
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
constantly==15.1.0
cryptography==3.2.1
cssselect2==0.4.1
daphne==3.0.1
defusedxml==0.7.0rc1
Django==3.1.2
django-braces==1.14.0
django-embed-video==1.3.3
django-extensions==3.0.9
django-localflavor==3.0.1
django-memcache-status==2.2
django-parler==2.2
django-rosetta==0.9.5
django-summernote==0.8.11.6
django-taggit==1.3.0
djangorestframework==3.12.2
easy-thumbnails==2.7
flower==0.9.5
gunicorn==20.0.4
hiredis==1.1.0
html5lib==1.1
humanize==3.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
kombu==5.0.2
Markdown==3.3.3
msgpack==1.0.0
oauthlib==3.1.0
Pillow==8.0.1
polib==1.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
Pyphen==0.10.0
python-memcached==1.59
python-stdnum==1.14
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
tinycss2==1.1.0
tornado==6.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.26.2
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==52.1
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
zope.interface==5.2.0
